Factor out radio_button.rs
